split tests into install and post-install projects
