+++ tests/integration/tests/shared/appWindow.spec.ts
@@ -1,6 +1,6 @@
 import { expect } from '@playwright/test';
 
-import { test } from './testExtensions';
+import { test } from '../../testExtensions';
 
 test('App window has title', async ({ app }) => {
   const window = await app.firstWindow();
